Fixed signed/unsigned mismatches.

+++ test/anvil/region_header_test.cpp
@@ -47,7 +47,7 @@ TEST(RegionHeader, CopyConstructor)
 
     // Init A
     amc::RegionHeader regionHeaderA(*info);
-    ASSERT_EQ(amc::ChunkCount, regionHeaderA.getRegionCount());
+    ASSERT_EQ(amc::ChunkCount, static_cast<int>(regionHeaderA.getRegionCount()));
 
     // Test Copy
     amc::RegionHeader regionHeaderB(regionHeaderA);
@@ -57,18 +57,18 @@ TEST(RegionHeader, CopyConstructor)
         EXPECT_EQ((*info)[i].getTimestamp(),   regionHeaderB.getChunkInfoAt(i).getTimestamp());
         EXPECT_EQ((*info)[i].getCompression(), regionHeaderB.getChunkInfoAt(i).getCompression());
     }
-    ASSERT_EQ(amc::ChunkCount, regionHeaderB.getRegionCount());
+    ASSERT_EQ(amc::ChunkCount, static_cast<int>(regionHeaderB.getRegionCount()));
 
     // Test empty copy
     amc::RegionHeader regionHeaderC;
     amc::RegionHeader regionHeaderD(regionHeaderC);
     for(int i = 0; i < amc::ChunkCount; ++i) {
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getOffset());
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getLength());
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getTimestamp());
+        EXPECT_EQ(0, static_cast<int>(regionHeaderD.getChunkInfoAt(i).getOffset()));
+        EXPECT_EQ(0, static_cast<int>(regionHeaderD.getChunkInfoAt(i).getLength()));
+        EXPECT_EQ(0, static_cast<int>(regionHeaderD.getChunkInfoAt(i).getTimestamp()));
         EXPECT_EQ(amc::CompressionType::GZip, regionHeaderD.getChunkInfoAt(i).getCompression());
     }
-    ASSERT_EQ(0, regionHeaderD.getRegionCount());
+    ASSERT_EQ(0, static_cast<int>(regionHeaderD.getRegionCount()));
 
     delete info;
 }
@@ -83,7 +83,7 @@ TEST(RegionHeader, MoveConstructor)
 
     // Init A
     amc::RegionHeader regionHeaderA(*info);
-    ASSERT_EQ(amc::ChunkCount, regionHeaderA.getRegionCount());
+    ASSERT_EQ(amc::ChunkCount, static_cast<int>(regionHeaderA.getRegionCount()));
 
     // Test Move
     amc::RegionHeader regionHeaderB(std::move(regionHeaderA));
@@ -93,18 +93,18 @@ TEST(RegionHeader, MoveConstructor)
         EXPECT_EQ((*info)[i].getTimestamp(),   regionHeaderB.getChunkInfoAt(i).getTimestamp());
         EXPECT_EQ((*info)[i].getCompression(), regionHeaderB.getChunkInfoAt(i).getCompression());
     }
-    ASSERT_EQ(amc::ChunkCount, regionHeaderB.getRegionCount());
+    ASSERT_EQ(amc::ChunkCount, static_cast<int>(regionHeaderB.getRegionCount()));
 
     // Test empty move
     amc::RegionHeader regionHeaderC;
     amc::RegionHeader regionHeaderD(std::move(regionHeaderC));
     for(int i = 0; i < amc::ChunkCount; ++i) {
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getOffset());
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getLength());
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getTimestamp());
+        EXPECT_EQ(0u, regionHeaderD.getChunkInfoAt(i).getOffset());
+        EXPECT_EQ(0u, regionHeaderD.getChunkInfoAt(i).getLength());
+        EXPECT_EQ(0u, regionHeaderD.getChunkInfoAt(i).getTimestamp());
         EXPECT_EQ(amc::CompressionType::GZip, regionHeaderD.getChunkInfoAt(i).getCompression());
     }
-    ASSERT_EQ(0, regionHeaderD.getRegionCount());
+    ASSERT_EQ(0, static_cast<int>(regionHeaderD.getRegionCount()));
 
     delete info;
 }
@@ -119,7 +119,7 @@ TEST(RegionHeader, CopyAssignment)
 
     // Init A
     amc::RegionHeader regionHeaderA(*info);
-    ASSERT_EQ(amc::ChunkCount, regionHeaderA.getRegionCount());
+    ASSERT_EQ(amc::ChunkCount, static_cast<int>(regionHeaderA.getRegionCount()));
 
     // Test Copy assignment
     amc::RegionHeader regionHeaderB = regionHeaderA;
@@ -129,18 +129,18 @@ TEST(RegionHeader, CopyAssignment)
         EXPECT_EQ((*info)[i].getTimestamp(),   regionHeaderB.getChunkInfoAt(i).getTimestamp());
         EXPECT_EQ((*info)[i].getCompression(), regionHeaderB.getChunkInfoAt(i).getCompression());
     }
-    ASSERT_EQ(amc::ChunkCount, regionHeaderB.getRegionCount());
+    ASSERT_EQ(amc::ChunkCount, static_cast<int>(regionHeaderB.getRegionCount()));
 
     // Test empty copy assignment
     amc::RegionHeader regionHeaderC;
     amc::RegionHeader regionHeaderD = regionHeaderC;
     for(int i = 0; i < amc::ChunkCount; ++i) {
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getOffset());
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getLength());
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getTimestamp());
+        EXPECT_EQ(0u, regionHeaderD.getChunkInfoAt(i).getOffset());
+        EXPECT_EQ(0u, regionHeaderD.getChunkInfoAt(i).getLength());
+        EXPECT_EQ(0u, regionHeaderD.getChunkInfoAt(i).getTimestamp());
         EXPECT_EQ(amc::CompressionType::GZip, regionHeaderD.getChunkInfoAt(i).getCompression());
     }
-    ASSERT_EQ(0, regionHeaderD.getRegionCount());
+    ASSERT_EQ(0, static_cast<int>(regionHeaderD.getRegionCount()));
 
     delete info;
 }
@@ -155,7 +155,7 @@ TEST(RegionHeader, MoveAssignment)
 
     // Init A
     amc::RegionHeader regionHeaderA(*info);
-    ASSERT_EQ(amc::ChunkCount, regionHeaderA.getRegionCount());
+    ASSERT_EQ(amc::ChunkCount, static_cast<int>(regionHeaderA.getRegionCount()));
 
     // Test Move assignment
     amc::RegionHeader regionHeaderB = std::move(regionHeaderA);
@@ -165,18 +165,18 @@ TEST(RegionHeader, MoveAssignment)
         EXPECT_EQ((*info)[i].getTimestamp(),   regionHeaderB.getChunkInfoAt(i).getTimestamp());
         EXPECT_EQ((*info)[i].getCompression(), regionHeaderB.getChunkInfoAt(i).getCompression());
     }
-    ASSERT_EQ(amc::ChunkCount, regionHeaderB.getRegionCount());
+    ASSERT_EQ(amc::ChunkCount, static_cast<int>(regionHeaderB.getRegionCount()));
 
     // Test empty move assignment
     amc::RegionHeader regionHeaderC;
     amc::RegionHeader regionHeaderD = std::move(regionHeaderC);
     for(int i = 0; i < amc::ChunkCount; ++i) {
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getOffset());
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getLength());
-        EXPECT_EQ(0, regionHeaderD.getChunkInfoAt(i).getTimestamp());
+        EXPECT_EQ(0u, regionHeaderD.getChunkInfoAt(i).getOffset());
+        EXPECT_EQ(0u, regionHeaderD.getChunkInfoAt(i).getLength());
+        EXPECT_EQ(0u, regionHeaderD.getChunkInfoAt(i).getTimestamp());
         EXPECT_EQ(amc::CompressionType::GZip, regionHeaderD.getChunkInfoAt(i).getCompression());
     }
-    ASSERT_EQ(0, regionHeaderD.getRegionCount());
+    ASSERT_EQ(0, static_cast<int>(regionHeaderD.getRegionCount()));
 
     delete info;
 }
@@ -208,10 +208,10 @@ TEST(RegionHeader, NotEqual)
 TEST(RegionHeader, getRegionCount)
 {
     amc::RegionHeader regionHeader;
-    ASSERT_EQ(0, regionHeader.getRegionCount());
+    ASSERT_EQ(0, static_cast<int>(regionHeader.getRegionCount()));
 
     regionHeader.setChunkInfoAt(1, amc::ChunkInfo(2, 2, 2, amc::CompressionType::GZip));
-    EXPECT_EQ(1, regionHeader.getRegionCount());
+    EXPECT_EQ(1, static_cast<int>(regionHeader.getRegionCount()));
 }
 
 TEST(RegionHeader, getRegionData)
@@ -235,7 +235,7 @@ TEST(RegionHeader, getChunkInfo)
         ASSERT_EQ((*info)[i].getTimestamp(),   regionHeader.getChunkInfoAt(i).getTimestamp());
         ASSERT_EQ((*info)[i].getCompression(), regionHeader.getChunkInfoAt(i).getCompression());
     }
-    ASSERT_EQ(amc::ChunkCount, regionHeader.getRegionCount());
+    ASSERT_EQ(amc::ChunkCount, static_cast<int>(regionHeader.getRegionCount()));
 
     // Check #2
     const std::array<amc::ChunkInfo, amc::ChunkCount> *info2 = &(regionHeader.getChunkInfo());
@@ -252,44 +252,44 @@ TEST(RegionHeader, getChunkInfo)
 TEST(RegionHeader, getChunkInfoAt)
 {
     amc::RegionHeader regionHeader;
-    ASSERT_EQ(0, regionHeader.getRegionCount());
+    ASSERT_EQ(0, static_cast<int>(regionHeader.getRegionCount()));
 
     amc::ChunkInfo info = regionHeader.getChunkInfoAt(10);
-    EXPECT_EQ(0, regionHeader.getChunkInfoAt(10).getOffset());
-    EXPECT_EQ(0, regionHeader.getChunkInfoAt(10).getLength());
-    EXPECT_EQ(0, regionHeader.getChunkInfoAt(10).getTimestamp());
+    EXPECT_EQ(0u, regionHeader.getChunkInfoAt(10).getOffset());
+    EXPECT_EQ(0u, regionHeader.getChunkInfoAt(10).getLength());
+    EXPECT_EQ(0u, regionHeader.getChunkInfoAt(10).getTimestamp());
     EXPECT_EQ(amc::CompressionType::GZip, regionHeader.getChunkInfoAt(10).getCompression());
 
     regionHeader.setChunkInfoAt(10, amc::ChunkInfo(2, 4, 6, amc::CompressionType::Zlib));
-    EXPECT_EQ(1, regionHeader.getRegionCount());
+    EXPECT_EQ(1, static_cast<int>(regionHeader.getRegionCount()));
 
     info = regionHeader.getChunkInfoAt(10);
-    EXPECT_EQ(2, regionHeader.getChunkInfoAt(10).getOffset());
-    EXPECT_EQ(4, regionHeader.getChunkInfoAt(10).getLength());
-    EXPECT_EQ(6, regionHeader.getChunkInfoAt(10).getTimestamp());
+    EXPECT_EQ(2u, regionHeader.getChunkInfoAt(10).getOffset());
+    EXPECT_EQ(4u, regionHeader.getChunkInfoAt(10).getLength());
+    EXPECT_EQ(6u, regionHeader.getChunkInfoAt(10).getTimestamp());
     EXPECT_EQ(amc::CompressionType::Zlib, regionHeader.getChunkInfoAt(10).getCompression());
 }
 
 TEST(RegionHeader, getChunkInfoAt_const)
 {
     amc::RegionHeader regionHeader;
-    ASSERT_EQ(0, regionHeader.getRegionCount());
+    ASSERT_EQ(0, static_cast<int>(regionHeader.getRegionCount()));
 
     const amc::RegionHeader regionHeaderB = regionHeader;
     amc::ChunkInfo info = regionHeaderB.getChunkInfoAt(10);
-    EXPECT_EQ(0, regionHeaderB.getChunkInfoAt(10).getOffset());
-    EXPECT_EQ(0, regionHeaderB.getChunkInfoAt(10).getLength());
-    EXPECT_EQ(0, regionHeaderB.getChunkInfoAt(10).getTimestamp());
+    EXPECT_EQ(0u, regionHeaderB.getChunkInfoAt(10).getOffset());
+    EXPECT_EQ(0u, regionHeaderB.getChunkInfoAt(10).getLength());
+    EXPECT_EQ(0u, regionHeaderB.getChunkInfoAt(10).getTimestamp());
     EXPECT_EQ(amc::CompressionType::GZip, regionHeaderB.getChunkInfoAt(10).getCompression());
 
     regionHeader.setChunkInfoAt(10, amc::ChunkInfo(2, 4, 6, amc::CompressionType::Zlib));
-    EXPECT_EQ(1, regionHeader.getRegionCount());
+    EXPECT_EQ(1, static_cast<int>(regionHeader.getRegionCount()));
 
     const amc::RegionHeader regionHeaderC = regionHeader;
     info = regionHeaderC.getChunkInfoAt(10);
-    EXPECT_EQ(2, regionHeaderC.getChunkInfoAt(10).getOffset());
-    EXPECT_EQ(4, regionHeaderC.getChunkInfoAt(10).getLength());
-    EXPECT_EQ(6, regionHeaderC.getChunkInfoAt(10).getTimestamp());
+    EXPECT_EQ(2u, regionHeaderC.getChunkInfoAt(10).getOffset());
+    EXPECT_EQ(4u, regionHeaderC.getChunkInfoAt(10).getLength());
+    EXPECT_EQ(6u, regionHeaderC.getChunkInfoAt(10).getTimestamp());
     EXPECT_EQ(amc::CompressionType::Zlib, regionHeaderC.getChunkInfoAt(10).getCompression());
 }
 
@@ -310,7 +310,7 @@ TEST(RegionHeader, setChunkInfo)
         ASSERT_EQ((*info)[i].getTimestamp(),   regionHeader.getChunkInfoAt(i).getTimestamp());
         ASSERT_EQ((*info)[i].getCompression(), regionHeader.getChunkInfoAt(i).getCompression());
     }
-    ASSERT_EQ(amc::ChunkCount, regionHeader.getRegionCount());
+    ASSERT_EQ(amc::ChunkCount, static_cast<int>(regionHeader.getRegionCount()));
 
     // Check #2
     const std::array<amc::ChunkInfo, amc::ChunkCount> *info2 = &(regionHeader.getChunkInfo());
@@ -327,19 +327,19 @@ TEST(RegionHeader, setChunkInfo)
 TEST(RegionHeader, setChunkInfoAt)
 {
     amc::RegionHeader regionHeader;
-    ASSERT_EQ(0, regionHeader.getRegionCount());
+    ASSERT_EQ(0u, regionHeader.getRegionCount());
 
     regionHeader.setChunkInfoAt(10, amc::ChunkInfo(2, 4, 6, amc::CompressionType::Zlib));
-    EXPECT_EQ(2, regionHeader.getChunkInfoAt(10).getOffset());
-    EXPECT_EQ(4, regionHeader.getChunkInfoAt(10).getLength());
-    EXPECT_EQ(6, regionHeader.getChunkInfoAt(10).getTimestamp());
+    EXPECT_EQ(2u, regionHeader.getChunkInfoAt(10).getOffset());
+    EXPECT_EQ(4u, regionHeader.getChunkInfoAt(10).getLength());
+    EXPECT_EQ(6u, regionHeader.getChunkInfoAt(10).getTimestamp());
     EXPECT_EQ(amc::CompressionType::Zlib, regionHeader.getChunkInfoAt(10).getCompression());
-    EXPECT_EQ(1, regionHeader.getRegionCount());
+    EXPECT_EQ(1, static_cast<int>(regionHeader.getRegionCount()));
 
     regionHeader.setChunkInfoAt(20, amc::ChunkInfo(4, 8, 12, amc::CompressionType::Uncompressed));
-    EXPECT_EQ(4, regionHeader.getChunkInfoAt(20).getOffset());
-    EXPECT_EQ(8, regionHeader.getChunkInfoAt(20).getLength());
-    EXPECT_EQ(12, regionHeader.getChunkInfoAt(20).getTimestamp());
+    EXPECT_EQ(4u, regionHeader.getChunkInfoAt(20).getOffset());
+    EXPECT_EQ(8u, regionHeader.getChunkInfoAt(20).getLength());
+    EXPECT_EQ(12u, regionHeader.getChunkInfoAt(20).getTimestamp());
     EXPECT_EQ(amc::CompressionType::Uncompressed, regionHeader.getChunkInfoAt(20).getCompression());
-    EXPECT_EQ(2, regionHeader.getRegionCount());
+    EXPECT_EQ(2, static_cast<int>(regionHeader.getRegionCount()));
 }
